refactor: extract route handlers into server_routes module
Move all axum handler fns and VERSION const out of server.rs into a new
server_routes.rs, leaving server.rs focused on binding, CORS, and routing.
