+++ src-tauri/src/server.rs
@@ -1,17 +1,15 @@
-use crate::adb_commands::{get_connected_device, get_device_serial, start_companion};
-use crate::settings::read_settings;
-use axum::extract::{Path, Request, State};
+use crate::server_routes::{
+    app_settings, device_connection_status, index, launch_companion_app_on_device, ping,
+};
+use axum::extract::Request;
 use axum::http::header::{CONTENT_TYPE, ORIGIN};
 use axum::http::{HeaderValue, Method};
-use axum::{Json, Router, ServiceExt, response::IntoResponse, routing::get};
-use serde_json::json;
+use axum::{Router, ServiceExt, routing::get};
 use tauri::AppHandle;
 use tokio::net::TcpListener;
 use tower::Layer;
 use tower_http::cors::CorsLayer;
 use tower_http::normalize_path::NormalizePathLayer;
-
-const VERSION: u32 = 2;
 
 pub(crate) async fn launch_server(app: AppHandle) {
     let router = Router::new()
@@ -26,13 +24,13 @@ pub(crate) async fn launch_server(app: AppHandle) {
         .layer(
             CorsLayer::new()
                 .allow_origin([
-                    "https://kodular.io".parse::<HeaderValue>().unwrap(),
-                    "https://starter.kodular.io".parse::<HeaderValue>().unwrap(),
-                    "https://creator.kodular.io".parse::<HeaderValue>().unwrap(),
-                    "https://c.kodular.io".parse::<HeaderValue>().unwrap(),
-                    "http://tauri.localhost".parse::<HeaderValue>().unwrap(),
-                    "tauri://localhost".parse::<HeaderValue>().unwrap(),
-                    "http://localhost:1420".parse::<HeaderValue>().unwrap(),
+                    HeaderValue::from_static("https://kodular.io"),
+                    HeaderValue::from_static("https://starter.kodular.io"),
+                    HeaderValue::from_static("https://creator.kodular.io"),
+                    HeaderValue::from_static("https://c.kodular.io"),
+                    HeaderValue::from_static("http://tauri.localhost"),
+                    HeaderValue::from_static("tauri://localhost"),
+                    HeaderValue::from_static("http://localhost:1420"),
                 ])
                 .allow_methods([Method::GET])
                 .allow_headers([ORIGIN, CONTENT_TYPE]),
@@ -50,35 +48,4 @@ pub(crate) async fn launch_server(app: AppHandle) {
     if let Err(e) = axum::serve(listener, ServiceExt::<Request>::into_make_service(app)).await {
         log::error!("HTTP server error: {e}");
     }
-}
-
-async fn index() -> impl IntoResponse {
-    "Hello World from Kodular Starter!"
-}
-
-async fn ping() -> impl IntoResponse {
-    Json(json!({ "status": "OK", "version": VERSION }))
-}
-
-async fn app_settings(State(app): State<AppHandle>) -> impl IntoResponse {
-    Json(read_settings(&app))
-}
-
-async fn device_connection_status(State(app): State<AppHandle>) -> impl IntoResponse {
-    let settings = read_settings(&app);
-    if let Some(mut device) = get_connected_device(&settings)
-        && let Some(serial) = get_device_serial(&mut device)
-    {
-        return Json(json!({ "status": "OK", "version": VERSION, "device": serial }));
-    }
-    Json(json!({ "status": "NO", "version": VERSION }))
-}
-
-async fn launch_companion_app_on_device(
-    State(app): State<AppHandle>,
-    Path(deviceid): Path<String>,
-) -> impl IntoResponse {
-    let settings = read_settings(&app);
-    let _ = start_companion(&deviceid, &settings);
-    "".into_response()
 }
